+++ src/pages/index.js
@@ -1,9 +1,9 @@
 export { default as AboutPage } from 'pages/about/About';
 export { default as ContactPage } from 'pages/contact/Contact';
 export { default as EntreMujeresPage } from 'pages/projects/EntreMujeresPage';
-export { default as ProjectsPage } from 'pages/projects/Projects';
 export { default as LandingPage } from 'pages/landing-page/LandingPage';
-export { default as RelatosVisualesPage } from 'pages/relatos-visuales/RelatosVisualesPage';
-export { default as RelatosVisualesDetailsPage } from 'pages/relatos-visuales/RelatosVisualesDetailsPage';
-export { default as RelatosVisualesListPage } from 'pages/relatos-visuales/RelatosVisualesListPage';
+export { default as PostDetailsPage } from 'pages/bitacora/PostDetailsPage';
+export { default as PostsListPage } from 'pages/bitacora/PostsListPage';
+export { default as ProjectsPage } from 'pages/projects/Projects';
+export { default as RelatosVisualesPage } from 'pages/projects/RelatosVisualesPage';
 export { default as TallerAdentroPage } from 'pages/projects/TallerAdentro';