+++ changelog.md
@@ -2,6 +2,7 @@
 
 -   Update: Change name of relatos visuales and remove link to blog.
 -   Feature: Move relatos visuales to projects and style it like the other ones.
+-   Feature: Change posts pages names to bitacora and link them on menu.
 
 ### v0.4.0